refactor: split core::llm into prompt / feedback / provider submodules
Roborev flagged the core::llm public surface (>5 names) on two consecutive
green commits (AC1, AC2); per the second-occurrence-promotes rule, escalate
to a structural split rather than defer a third time. The pure prompt core
(constants, Submission, ExecResults, Prompt, build_prompt) moves to
prompt.rs; the effectful request (Verdict, Feedback DTO, FeedbackError,
request_feedback) to feedback.rs; ProviderChoice stays in provider.rs. This
also realizes the §2.1/§2.2 pure-vs-effectful cut the module doc claimed.

The public names are re-exported from mod.rs, so callers keep using
core::llm::build_prompt, core::llm::OPEN_CODE, etc. — the spec's "exported
from core::llm" requirement holds and the integration test is unchanged.
No behavior change; 56/56 green, clippy clean.
