+++ crates/core/src/llm/mod.rs
@@ -1,531 +1,25 @@
-//! The LLM feedback layer: a pure, injection-hardened prompt builder, and (in
-//! later slices) the rig-core provider client that turns a model's structured
-//! tool call into a typed verdict.
+//! The LLM feedback layer: a pure prompt builder and the rig-core provider client
+//! that turns a model's structured tool call into a typed verdict.
 //!
-//! This module owns prompt construction, provider-client construction, and the
-//! boundary DTO → domain mapping (ADR-0006). It does **not** run learner code or
+//! Three concerns, one submodule each (§4.1): [`prompt`] builds the
+//! injection-hardened prompt (pure, §2.1); [`provider`] is the closed set of LLM
+//! backends; [`feedback`] runs the request and maps the result to a [`Verdict`]
+//! (the only effectful part, §2.2). The split keeps the pure prompt core separate
+//! from the effectful request (§3.1). This layer owns prompt-build, provider-client
+//! construction, and the DTO → domain mapping; it does **not** run learner code or
 //! score evals — that is the [`runner`](crate::runner) and [`grade`](crate::grade)
-//! layers' job (§4.1). The student submission is untrusted input bound for an
-//! LLM, so [`build_prompt`] neutralizes every interpolated value: a forged
-//! delimiter or label can never become a second structural token.
+//! layers' job. See ADR-0006.
+//!
+//! The public names are re-exported here, so callers use `core::llm::build_prompt`,
+//! `core::llm::OPEN_CODE`, etc., without depending on the internal split.
 
-use std::fmt;
-
-use rig_core::prelude::CompletionClient;
-use rig_core::providers::{anthropic, openai};
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-
-use crate::grade::CheckOutcome;
-use crate::lesson::Lesson;
-use crate::runner::ExecutionResult;
-
+mod feedback;
+mod prompt;
 mod provider;
 
+pub use feedback::{FeedbackError, Verdict, request_feedback};
+pub use prompt::{
+    CHECKS_LABEL, CLOSE_CODE, ExecResults, OPEN_CODE, OUTPUT_LABEL, Prompt, Submission,
+    build_prompt,
+};
 pub use provider::ProviderChoice;
-
-/// Opens the fence around the verbatim student submission.
-///
-/// Exported so tests and the builder reference one literal source — the same
-/// constant is both emitted by [`build_prompt`] and asserted against, so the
-/// test predicate pins the real delimiter rather than a hand-copied twin that
-/// could drift from it (§1.5).
-pub const OPEN_CODE: &str = "<<<STUDENT_CODE_BEGIN>>>";
-/// Closes the fence around the verbatim student submission.
-pub const CLOSE_CODE: &str = "<<<STUDENT_CODE_END>>>";
-/// Labels the captured-output section.
-pub const OUTPUT_LABEL: &str = "<<<CAPTURED_OUTPUT>>>";
-/// Labels the check-results section.
-pub const CHECKS_LABEL: &str = "<<<CHECK_RESULTS>>>";
-
-/// What replaces any structural token found inside untrusted interpolated text,
-/// so an injected fence or label can never count as a real one.
-const NEUTRALIZED: &str = "[neutralized-delimiter]";
-
-/// A learner's submitted code, awaiting feedback.
-///
-/// A newtype around the source so submitted code is never confused with arbitrary
-/// text elsewhere (§1.4).
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Submission {
-    /// The submitted source code, verbatim.
-    pub code: String,
-}
-
-impl Submission {
-    /// Wrap submitted source `code`.
-    pub fn new(code: impl Into<String>) -> Self {
-        Self { code: code.into() }
-    }
-}
-
-/// A graded run: what the submission produced, paired with its per-check verdicts.
-///
-/// Bundles the single [`ExecutionResult`] from running the submission with the
-/// ordered [`CheckOutcome`]s from grading it — one per lesson check, by index. It
-/// does **not** carry the check code-strings; those live on the [`Lesson`] and are
-/// paired with `outcomes` positionally (ADR-0006). This is the input
-/// [`build_prompt`] reads.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ExecResults {
-    /// What running the submission produced.
-    pub output: ExecutionResult,
-    /// One verdict per lesson check, in `lesson.checks` order.
-    pub outcomes: Vec<CheckOutcome>,
-}
-
-/// A rendered LLM feedback prompt.
-///
-/// A newtype around the assembled text so a prompt is never confused with
-/// arbitrary text, and the rendering stays the single thing the provider sends
-/// (§1.4).
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Prompt(String);
-
-impl Prompt {
-    /// The rendered prompt text.
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-/// Strip every structural token out of untrusted `text`.
-///
-/// Applied to each interpolated value so the only occurrences of the fences and
-/// labels in the final prompt are the ones [`build_prompt`] itself emits. The
-/// replacement marker contains no structural token, so replacements can neither
-/// create nor hide one another. This is the whole injection defense: an injected
-/// `CLOSE_CODE` + forged `CHECKS_LABEL` cannot break out of the fence or forge a
-/// verdict, because each is rewritten before it reaches the prompt.
-fn neutralize(text: &str) -> String {
-    text.replace(OPEN_CODE, NEUTRALIZED)
-        .replace(CLOSE_CODE, NEUTRALIZED)
-        .replace(OUTPUT_LABEL, NEUTRALIZED)
-        .replace(CHECKS_LABEL, NEUTRALIZED)
-}
-
-/// Render one check as `<check code>: <outcome>` for the checks section, both
-/// sides neutralized.
-fn render_check(check: &str, outcome: &CheckOutcome) -> String {
-    let verdict = match outcome {
-        CheckOutcome::Pass => "pass".to_string(),
-        CheckOutcome::Fail { detail } => format!("fail — {}", neutralize(detail)),
-        CheckOutcome::NotRun { reason } => format!("not run — {}", neutralize(reason)),
-    };
-    format!("{}: {verdict}", neutralize(check))
-}
-
-/// Render the check-results section: one line per outcome, in order, each labeled
-/// with its lesson check code-string.
-///
-/// Iterates over the **outcomes** — the verdicts — not the lesson's checks, so a
-/// verdict is never silently dropped. In the normal flow the two are 1:1:
-/// [`run_checks`](crate::grade::run_checks) returns exactly one outcome per
-/// `lesson.checks` entry, in order. If a caller hand-built [`ExecResults`] with
-/// more outcomes than the lesson has checks, the extra outcomes still render
-/// (labeled by 1-based position) rather than vanishing; checks with no outcome
-/// have no verdict to report and so contribute no line.
-fn render_checks(lesson: &Lesson, outcomes: &[CheckOutcome]) -> String {
-    outcomes
-        .iter()
-        .enumerate()
-        .map(|(i, outcome)| {
-            let fallback = format!("check {}", i + 1);
-            let label = lesson
-                .checks
-                .get(i)
-                .map(String::as_str)
-                .unwrap_or(fallback.as_str());
-            render_check(label, outcome)
-        })
-        .collect::<Vec<_>>()
-        .join("\n")
-}
-
-/// Build the LLM feedback prompt for a graded submission.
-///
-/// Pure (§2.1): it reads only the borrowed domain values and performs no IO, env
-/// read, or network call, so identical inputs always render byte-identically. The
-/// layout is a fixed structure — the task, a single fenced copy of the submission,
-/// a captured-output section, and a check-results section (one line per outcome,
-/// labeled with its `lesson.checks` entry by index) — not the lesson's
-/// `llm_evaluation_prompt` template (ADR-0006). `results.outcomes` is expected 1:1
-/// with `lesson.checks` (the shape `run_checks` produces); the rendering never
-/// silently drops a verdict if they desync (see [`render_checks`]). Every
-/// interpolated value is [`neutralize`]d, so the fences and labels each appear
-/// exactly once even when the submission forges them: injected text can never be
-/// read as code or as a verdict.
-pub fn build_prompt(lesson: &Lesson, submission: &Submission, results: &ExecResults) -> Prompt {
-    let task = neutralize(&lesson.exercise.prompt);
-    let code = neutralize(&submission.code);
-    let output = neutralize(&results.output.stdout);
-    let checks = render_checks(lesson, &results.outcomes);
-
-    let rendered = [
-        "You are evaluating student code for a programming exercise.",
-        "",
-        "Task:",
-        task.trim_end(),
-        "",
-        OPEN_CODE,
-        code.trim_end_matches('\n'),
-        CLOSE_CODE,
-        "",
-        OUTPUT_LABEL,
-        output.trim_end_matches('\n'),
-        "",
-        CHECKS_LABEL,
-        &checks,
-    ]
-    .join("\n");
-
-    Prompt(rendered)
-}
-
-/// A graded verdict on a submission: correct or incorrect, each carrying the
-/// learner-facing message.
-///
-/// A sum type, not a `bool` plus a message, so "correct with no message" and
-/// contradictory states are unrepresentable and there is no public `is_correct`
-/// (§1.2). This is what callers consume; the rig DTO ([`Feedback`]) stays an
-/// implementation detail of this module.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum Verdict {
-    /// The submission satisfies the exercise.
-    Correct {
-        /// The learner-facing feedback message.
-        message: String,
-    },
-    /// The submission does not satisfy the exercise.
-    Incorrect {
-        /// The learner-facing feedback message.
-        message: String,
-    },
-}
-
-/// The structured feedback the model returns through its tool call — the boundary
-/// DTO rig's `Extractor` fills from the tool-call arguments (ADR-0006).
-///
-/// `is_correct` is a non-`Option` `bool` and the struct derives **no** `Default`,
-/// so a tool call that omits a field is a deserialize error, never a
-/// silently-defaulted grade. It is mapped into [`Verdict`] at the boundary so the
-/// bool's meaning lives in the variant; callers never see this type.
-#[derive(Debug, Deserialize, Serialize, JsonSchema)]
-struct Feedback {
-    /// Whether the submission satisfies the exercise.
-    is_correct: bool,
-    /// The feedback message for the learner.
-    feedback_message: String,
-}
-
-impl From<Feedback> for Verdict {
-    fn from(feedback: Feedback) -> Self {
-        let Feedback {
-            is_correct,
-            feedback_message,
-        } = feedback;
-        if is_correct {
-            Verdict::Correct {
-                message: feedback_message,
-            }
-        } else {
-            Verdict::Incorrect {
-                message: feedback_message,
-            }
-        }
-    }
-}
-
-/// Why feedback could not be produced.
-///
-/// A typed error implementing [`std::error::Error`], so `core` stays free of
-/// `anyhow` (ADR-0001); the CLI maps it at the edge.
-#[derive(Debug)]
-pub enum FeedbackError {
-    /// The provider client could not be constructed.
-    Client(String),
-    /// The completion request itself failed — a transport, auth, or model error
-    /// (e.g. the provider rejected the key, or the network was unreachable). The
-    /// caller's recourse is to retry or fix credentials, not to re-prompt.
-    Completion(String),
-    /// The model responded, but its output could not be turned into a verdict — a
-    /// malformed or missing field, or no tool call at all. The model's output, not
-    /// the transport, is at fault.
-    Extraction(String),
-}
-
-impl fmt::Display for FeedbackError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            FeedbackError::Client(msg) => write!(f, "could not build the provider client: {msg}"),
-            FeedbackError::Completion(msg) => write!(f, "the feedback request failed: {msg}"),
-            FeedbackError::Extraction(msg) => {
-                write!(
-                    f,
-                    "could not extract a verdict from the model response: {msg}"
-                )
-            }
-        }
-    }
-}
-
-impl std::error::Error for FeedbackError {}
-
-impl From<rig_core::extractor::ExtractionError> for FeedbackError {
-    /// Preserve rig's own distinction between a failed request and an
-    /// unparseable response (§1.2): a `CompletionError` is a transport/auth/model
-    /// failure, while a deserialize failure or missing tool call means the
-    /// response could not be turned into a verdict.
-    fn from(error: rig_core::extractor::ExtractionError) -> Self {
-        use rig_core::extractor::ExtractionError;
-        match error {
-            ExtractionError::CompletionError(e) => FeedbackError::Completion(e.to_string()),
-            ExtractionError::DeserializationError(e) => FeedbackError::Extraction(e.to_string()),
-            ExtractionError::NoData => {
-                FeedbackError::Extraction("the model returned no structured feedback".to_string())
-            }
-        }
-    }
-}
-
-/// Request structured feedback for a `prompt` from the chosen provider.
-///
-/// The effectful counterpart to [`build_prompt`] (§2.2): it reads the active
-/// provider's API key, builds the rig client, runs an `Extractor<Feedback>` over
-/// the prompt, and maps the extracted DTO into a [`Verdict`]. Fireworks is reached
-/// through rig's OpenAI **chat-completions** client (Fireworks is
-/// OpenAI-compatible); Anthropic through its native client. `base_url_override`
-/// points the client at a mock server in tests — the test seam (ADR-0006);
-/// production passes `None`. Callers depend only on this function and [`Verdict`],
-/// never on rig types.
-pub async fn request_feedback(
-    provider: ProviderChoice,
-    prompt: &Prompt,
-    base_url_override: Option<&str>,
-) -> Result<Verdict, FeedbackError> {
-    let key = std::env::var(provider.key_var()).unwrap_or_default();
-    let base_url = base_url_override.unwrap_or(provider.default_base_url());
-    let model = provider.default_model();
-
-    let extracted = match provider {
-        ProviderChoice::Fireworks => {
-            let client = openai::Client::builder()
-                .api_key(key)
-                .base_url(base_url)
-                .build()
-                .map_err(|e| FeedbackError::Client(e.to_string()))?;
-            client
-                .completions_api()
-                .extractor::<Feedback>(model)
-                .build()
-                .extract(prompt.as_str())
-                .await
-        }
-        ProviderChoice::Anthropic => {
-            // TODO(anthropic-happy-path): the Anthropic success path has no
-            // wiremock coverage — its native messages-API response shape differs
-            // from the OpenAI envelope `mount_tool_call` builds, so it needs its
-            // own tool_use-block fixture. AC3 exercises only the Anthropic key
-            // guard. See docs/agent-notes/llm.md.
-            let client = anthropic::Client::builder()
-                .api_key(key)
-                .base_url(base_url)
-                .build()
-                .map_err(|e| FeedbackError::Client(e.to_string()))?;
-            client
-                .extractor::<Feedback>(model)
-                .build()
-                .extract(prompt.as_str())
-                .await
-        }
-    };
-
-    extracted.map(Verdict::from).map_err(FeedbackError::from)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn neutralize_rewrites_every_structural_token() {
-        for token in [OPEN_CODE, CLOSE_CODE, OUTPUT_LABEL, CHECKS_LABEL] {
-            let injected = format!("before {token} after");
-            let cleaned = neutralize(&injected);
-            assert!(
-                !cleaned.contains(token),
-                "neutralize must rewrite {token}, got: {cleaned}"
-            );
-            assert!(
-                cleaned.contains(NEUTRALIZED),
-                "the token is replaced by the neutral marker"
-            );
-        }
-    }
-
-    #[test]
-    fn neutralize_marker_introduces_no_structural_token() {
-        // The marker must itself be free of every structural token, or a
-        // replacement could resurrect one it just removed.
-        for token in [OPEN_CODE, CLOSE_CODE, OUTPUT_LABEL, CHECKS_LABEL] {
-            assert!(
-                !NEUTRALIZED.contains(token),
-                "the neutral marker must not contain {token}"
-            );
-        }
-    }
-
-    #[test]
-    fn render_check_distinguishes_the_three_outcomes() {
-        assert_eq!(render_check("c", &CheckOutcome::Pass), "c: pass");
-        assert_eq!(
-            render_check(
-                "c",
-                &CheckOutcome::Fail {
-                    detail: "boom".to_string()
-                }
-            ),
-            "c: fail — boom"
-        );
-        assert_eq!(
-            render_check(
-                "c",
-                &CheckOutcome::NotRun {
-                    reason: "no run".to_string()
-                }
-            ),
-            "c: not run — no run"
-        );
-    }
-
-    fn lesson_with_checks(checks: &[&str]) -> Lesson {
-        let check_lines = checks
-            .iter()
-            .map(|c| format!("  - \"{c}\""))
-            .collect::<Vec<_>>()
-            .join("\n");
-        let yaml = format!(
-            "lesson_name: \"L\"\nlanguage: R\nchecks:\n{check_lines}\nexercise:\n  \
-             prompt: \"do it\"\n  llm_evaluation_prompt: \"grade {{student_code}}\"\n"
-        );
-        Lesson::parse(&yaml).expect("the constructed lesson is valid")
-    }
-
-    #[test]
-    fn render_checks_labels_each_outcome_with_its_check_in_order() {
-        let lesson = lesson_with_checks(&["first_check", "second_check"]);
-        let rendered = render_checks(
-            &lesson,
-            &[
-                CheckOutcome::Pass,
-                CheckOutcome::Fail {
-                    detail: "nope".to_string(),
-                },
-            ],
-        );
-        assert_eq!(rendered, "first_check: pass\nsecond_check: fail — nope");
-    }
-
-    #[test]
-    fn render_checks_never_drops_a_verdict_when_outcomes_exceed_checks() {
-        // A desynced ExecResults (more outcomes than the lesson has checks) must
-        // not hide a verdict — the security-relevant property is that a Fail is
-        // always shown. The unlabeled extra falls back to its 1-based position.
-        let lesson = lesson_with_checks(&["only_check"]);
-        let rendered = render_checks(
-            &lesson,
-            &[
-                CheckOutcome::Pass,
-                CheckOutcome::Fail {
-                    detail: "hidden?".to_string(),
-                },
-            ],
-        );
-        assert_eq!(rendered, "only_check: pass\ncheck 2: fail — hidden?");
-        assert!(
-            rendered.contains("hidden?"),
-            "the second verdict must not be dropped, got: {rendered}"
-        );
-    }
-
-    #[test]
-    fn render_checks_renders_no_line_for_a_check_without_an_outcome() {
-        // The other desync direction: more checks than outcomes. A check with no
-        // outcome has no verdict to report, so it contributes no line — it must
-        // never surface as a phantom pass.
-        let lesson = lesson_with_checks(&["graded", "ungraded"]);
-        let rendered = render_checks(&lesson, &[CheckOutcome::Pass]);
-        assert_eq!(rendered, "graded: pass");
-        assert!(
-            !rendered.contains("ungraded"),
-            "a check with no outcome must not appear as a phantom verdict, got: {rendered}"
-        );
-    }
-
-    #[test]
-    fn render_check_neutralizes_a_token_in_the_outcome_detail() {
-        // A check's failure detail is interpreter output — untrusted — so a fence
-        // smuggled through stderr must be neutralized too.
-        let rendered = render_check(
-            "c",
-            &CheckOutcome::Fail {
-                detail: format!("oops {CLOSE_CODE}"),
-            },
-        );
-        assert!(
-            !rendered.contains(CLOSE_CODE),
-            "a token in a check detail is neutralized, got: {rendered}"
-        );
-    }
-
-    #[test]
-    fn feedback_maps_to_the_matching_verdict_variant() {
-        // The bool's meaning lives in the variant: is_correct true → Correct,
-        // false → Incorrect, and the message passes through unchanged.
-        let correct: Verdict = Feedback {
-            is_correct: true,
-            feedback_message: "well done".to_string(),
-        }
-        .into();
-        assert_eq!(
-            correct,
-            Verdict::Correct {
-                message: "well done".to_string()
-            }
-        );
-
-        let incorrect: Verdict = Feedback {
-            is_correct: false,
-            feedback_message: "try again".to_string(),
-        }
-        .into();
-        assert_eq!(
-            incorrect,
-            Verdict::Incorrect {
-                message: "try again".to_string()
-            }
-        );
-    }
-
-    #[test]
-    fn extraction_failures_map_to_the_extraction_kind() {
-        use rig_core::extractor::ExtractionError;
-
-        // No tool call at all → an Extraction error: the response, not transport,
-        // failed to yield a verdict.
-        let no_data: FeedbackError = ExtractionError::NoData.into();
-        assert!(
-            matches!(no_data, FeedbackError::Extraction(_)),
-            "NoData is an extraction failure, got {no_data:?}"
-        );
-
-        // A malformed/missing field → an Extraction error, kept distinct from a
-        // transport-level Completion failure (§1.2).
-        let deser_err = serde_json::from_str::<i32>("not-an-int").unwrap_err();
-        let bad_field: FeedbackError = ExtractionError::DeserializationError(deser_err).into();
-        assert!(
-            matches!(bad_field, FeedbackError::Extraction(_)),
-            "a deserialize failure is an extraction failure, got {bad_field:?}"
-        );
-    }
-}
